From: Kurt Taylor <krtaylor@us.ibm.com>
Attached is a patch for Mantis 25 - this fixes the problem of not having a
touch_start happen for all prims in a linked group.  So, with this, large
builds can now have a single script in the base prim and it will run when
any prim in the linked build is touched.   The problem was that the
objectgrab event was not being propagated to all the prims in the group.
 


git-svn-id: svn+ssh://opensimulator.org/var/svn/opensim/trunk@3341 e9a5c24e-b13f-11db-933b-17aa6df88f97