i can haz pantz?
You sure can.  This change set restores pants (and the rest of the
default appearance) in grid mode.  The 
root issue had to do with serializing multi-faced textures to the 
grid server.  This also restores the lookup path through the avatar
factory module, as that seems the reasonable place to have it live.
Some clean up patches are coming later as well, plus testing on 
standalone, but this should be in a good kicking around state for 
grid users.



git-svn-id: svn+ssh://opensimulator.org/var/svn/opensim/trunk@4759 e9a5c24e-b13f-11db-933b-17aa6df88f97