From: Kurt Taylor <krtaylor@us.ibm.com>
Attached is an initial implementation of llGetNotecardLine and
llGetNumberOfNotecardLines. I decided to go ahead an send these out for
comment while I continue to work on the second part of the proper
implementation. These functions work and return the values requested, as
initially defined in the code, but should be properly implemented to return
the requested information via a dataserver event.  This
event will be added and these functions fixed and included in a second
patch shortly.


git-svn-id: svn+ssh://opensimulator.org/var/svn/opensim/trunk@4893 e9a5c24e-b13f-11db-933b-17aa6df88f97