From: Christopher Yeoh <cyeoh@au1.ibm.com>
The problem I noticed is that when child prims are selected unlike
the Linden server, OpenSim does not return property information - it
only searches the root prims for matches.

What I'm not sure about is whether the parcel prim count taint should
be set or if the root prim should be set as selected (doesn't
seem possible to set the child prim as selected). It does
however fix the problem of the property information
not being sent, so its an improvement (I think!).


git-svn-id: svn+ssh://opensimulator.org/var/svn/opensim/trunk@5619 e9a5c24e-b13f-11db-933b-17aa6df88f97