From: Richard Alimi <ralimi@us.ibm.com>
The following is a patch that causes the ensuing http_response event (after
an llHTTPRequest) to include the HTTP status code returned from the server
(if available). The patch also sets the body parameter for the
http_response event to be set as the status description returned by the
server.



git-svn-id: svn+ssh://opensimulator.org/var/svn/opensim/trunk@5712 e9a5c24e-b13f-11db-933b-17aa6df88f97