From: alan webb <alan_webb@us.ibm.com>
This fixes a sitting problem which popped up on loading regions from
archive and turned out to be caused by the fact that the archiving
mechanisms all preserve the sit-target information that is set in the
object at the time the image is captured. This caused the new region
to pick a sit-target prim that did not correspond to the prim on which
the script was running, so the script's changed event is driven with
an invalid avatar UUID.

I have modified the ArchiveReadRequest class so that any sit information
that survives the archiving process is deleted before the object is
instantiated. This change has just been checked in.



git-svn-id: svn+ssh://opensimulator.org/var/svn/opensim/trunk@5772 e9a5c24e-b13f-11db-933b-17aa6df88f97