From: mike pitman <pitman@us.ibm.com>
fixes the terrain spikes, and is the result of mostly a tuning
operation on the smooth and flatten tools. I dug in and found that the
spikes apparently result from smooth's overly aggressive iteration
steps toward the average curvature, which leads to an instability that
blows up the heights. I introduced a scaling factor to dampen the
'duration' parameter which tames progress and seems to keep things
stable.



git-svn-id: svn+ssh://opensimulator.org/var/svn/opensim/trunk@6302 e9a5c24e-b13f-11db-933b-17aa6df88f97