From: Christopher Yeoh <yeohc@au1.ibm.com>
The attached patch ads an OpenSim.ini option (AutomaticLinkPermission)
which when enabled makes PERMISSION_CHANGE_LINKS to be granted to
scripts by default. When enabled llGetPermissions will always return it
as granted and llCreateLink//llBreakLink will succeed without
doing llRequestPermissions. ONLY ENABLE THIS IN TRUSTED ENVIRONMENTS.

The patch also fixes a minor bug in llCreateLink
related to a potential dereference of a null client object.

git-svn-id: svn+ssh://opensimulator.org/var/svn/opensim/trunk@7647 e9a5c24e-b13f-11db-933b-17aa6df88f97