From Rob Smart <SMARTROB@uk.ibm.com>
In SL if llAbs() is called with the minimum integer value of -2147483648
it will return that value untouched without error.

this patch replicates the SL functionality.

OpenSim currently throws an overflow exception: number too small under
mono or a "System.OverflowException: Negating the minimum value of a
twos complement number is invalid. " under .NET

git-svn-id: svn+ssh://opensimulator.org/var/svn/opensim/trunk@8317 e9a5c24e-b13f-11db-933b-17aa6df88f97