From: Christopher Yeoh <yeohc@au1.ibm.com>
This changeset add the RegionReady module code. The module sends a
message on a configurable channel when an oar file has finished
loading or if the script engine has emptied its queue for the first
time (eg server startup). Config is something like this:

[RegionReady]
enabled = true
channel_notify = -800

The module also knows if there was an error with startup.


git-svn-id: svn+ssh://opensimulator.org/var/svn/opensim/trunk@8335 e9a5c24e-b13f-11db-933b-17aa6df88f97