From: Christopher Yeoh <yeohc@au1.ibm.com>
Attached is a patch which enabled through an OpenSim.ini option the ability
to read long notecard lines. Currently although the data is read from
the notecard it is truncated at 255 characters (same as for the LL servers.
This patch allows the setting of that limit to a different value.

    ; Maximum length of notecard line read
    ; Increasing this to large values potentially opens
    ; up the system to malicious scripters
    ; NotecardLineReadCharsMax = 255

this allows for save/restore functionality using notecards without
having to worry about very short line length limits.

git-svn-id: svn+ssh://opensimulator.org/var/svn/opensim/trunk@8804 e9a5c24e-b13f-11db-933b-17aa6df88f97