From: Christopher Yeoh <yeohc@au1.ibm.com>
The attached patch implements osKey2Name and osName2Key
which converts between a UUID key for an avatar and
an avatar name and vice-versa.

osKey2Name is similar to llKey2Name except that it will work
even if the avatar being looked up is not in the same region
as the script.

git-svn-id: svn+ssh://opensimulator.org/var/svn/opensim/trunk@9071 e9a5c24e-b13f-11db-933b-17aa6df88f97